adds new files to split components
