+++ react-essentials/src/components/TabButton.jsx
@@ -1,0 +1,7 @@
+export default function TabButton({ children }) {
+  return (
+    <li>
+      <button>{children}</button>
+    </li>
+  );
+}
