+++ react-essentials/src/index.css
@@ -1,11 +1,11 @@
-@import url("https://fonts.googleapis.com/css2?family=Quicksand:wght@400;700&family=Roboto+Condensed:wght@400;700&display=swap");
+@import url('https://fonts.googleapis.com/css2?family=Quicksand:wght@400;700&family=Roboto+Condensed:wght@400;700&display=swap');
 
 * {
   box-sizing: border-box;
 }
 
 html {
-  font-family: "Quicksand", sans-serif;
+  font-family: 'Quicksand', sans-serif;
   line-height: 1.5;
 
   color: #ebe7ef;
@@ -18,35 +18,6 @@ html {
   -webkit-text-size-adjust: 100%;
   min-height: 80rem;
 }
-
-header {
-  text-align: center;
-  margin: 3rem 0;
-}
-
-header img {
-  height: 5rem;
-  width: 10rem;
-  object-fit: cover;
-}
-
-header h1 {
-  margin: 0;
-  font-family: "Roboto Condensed", sans-serif;
-  font-size: 5rem;
-  background: linear-gradient(40deg, #ea00ff, #ea00ff, #03d5ff, #03d5ff);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  filter: drop-shadow(0 2px 8px rgba(0, 0, 0, 0.5));
-}
-
-header p {
-  margin: 0;
-  font-size: 1.25rem;
-  color: #8964b0;
-  font-family: "Roboto Condensed", sans-serif;
-}
-
 main {
   width: 90%;
   max-width: 50rem;
@@ -62,7 +33,7 @@ main {
 
 main h2 {
   text-align: center;
-  font-family: "Roboto Condensed", sans-serif;
+  font-family: 'Roboto Condensed', sans-serif;
   margin: 0 0 1.5rem 0;
   color: #a18aba;
 }
@@ -118,7 +89,7 @@ main h2 {
   border-radius: 6px;
   background-color: transparent;
   color: #a18aba;
-  font-family: "Roboto Condensed", sans-serif;
+  font-family: 'Roboto Condensed', sans-serif;
   font-size: 1rem;
   cursor: pointer;
   transition: all 0.2s ease-in-out;
